pyproject.toml
[tool.poetry]
name = "astx"
version = "0.2.1"  # semantic-release
description = "OSL Python Package contains all the boilerplate you need to create a Python package."
authors = ["Ivan Ogasawara <ivan.ogasawara@gmail.com>"]
license = "BSD 3 Clause"
exclude = [
  ".git/*",
  ".env*",
]
packages = [
  {include = "astx", from="src"},
]
include = ["src/astx/py.typed"]

[tool.poetry.dependencies]
python = "^3.8.1"
atpublic = "^4.0"
typing-extensions = { version = "^4", python = "<3.9" }

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"
pytest-cov = "^4.1.0"
coverage = "^7.2.7"
black = "^23.3.0"
pre-commit = "^3.3.2"
ruff = "^0.0.278"
mypy = "^1.3.0"
bandit = "^1.7.5"
vulture = ">=2.9.1"
mccabe = "^0.6.1"
ipython = "<8"
ipykernel = ">=6.0.0"
mkdocs = "^1.4.3"
mkdocs-exclude = "^1.0.2"
mkdocs-jupyter = "^0.24.1"
mkdocs-literate-nav = "^0.6.0"
mkdocs-macros-plugin = ">=0.7.0,<1"
mkdocs-material = "^9.1.15"
mkdocstrings = "^0.21.2"
mkdocstrings-python = "^1.1.2"
makim = "1.8.3"

[tool.pytest.ini_options]
testpaths = [
    "tests",
]

[tool.black]
line-length = 79
target-version = ["py38"]
force-exclude = '''(?x)(
    docs/*
  | .*\\.egg-info
)'''  # TOML's single-quoted strings do not require escaping backslashes

[tool.bandit]
exclude_dirs = ["tests"]
targets = "./"

[tool.vulture]
exclude = ["tests"]
ignore_decorators = ["@abc.abstractmethod"]
ignore_names = []
make_whitelist = true
min_confidence = 80
paths = ["./"]
sort_by_size = true
verbose = false

[tool.ruff]
line-length = 79
force-exclude = true
src = ["./"]
exclude = [
  'docs',
]
select = [
  "E",   # pycodestyle
  "F",   # pyflakes
  "D",   # pydocstyle
  "YTT", # flake8-2020
  "PL",  # PL
  "RUF", # Ruff-specific rules
  "I001", # isort
]
fixable = ["I001"]

[tool.ruff.pydocstyle]
convention = "numpy"

[tool.ruff.isort]
# Use a single line between direct and from import
lines-between-types = 1

[tool.mypy]
python_version = "3.8"
check_untyped_defs = true
strict = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
